feat: adding context.h containing PyVeloxContext
